Add wasmldr integration test(s)
We can run wasm inside keeps now - let's prove it!

Here's what this patch does:

* Move wasm sources (*.wat) from internal/wasmldr/fixtures to tests/wasm
  * Copy .wat contents into wasmldr unittests and build inside the test
  * Remove wasmldr build.rs and update Cargo.toml
* Build tests/wasm/*.wat to OUT_DIR/bin/*.wasm in build.rs
* Add tests/wasmldr_tests.rs, which runs the .wasm test binaries

The intent is that we should be adding more wasm tests over time; this
gives us a place to put them and a harness to run them.

The tricky bit was getting Rust to pass the module to the
child process, since right now the only way we have to do that is to
pass it on FD3 - but Rust doesn't have `dup2()` and it *really* wants to
set FD_CLOEXEC on everything it opens, so we have to do some unsafe
things to actually pass FDs to a child process.

This isn't pretty, but it's a start.

Signed-off-by: Will Woods <will@profian.com>