Hide `enarx exec` and document it as being for test/debug only
`enarx exec` is really only useful for testing, debugging, and
development. We'd like to remove it from release builds, but then our
integration tests fail, because they use `enarx exec` to run test
binaries.

So, this commit marks `exec` with `AppSettings::Hidden` so it doesn't
show up in the toplevel `enarx` help output. It also expands the help
text to make it clear that `enarx exec` is an internal interface, only
intended for debugging and testing.

Also, just for clarity, rename the exec argument field to `binpath` and
show it as "BINARY" in the help text.

Signed-off-by: Will Woods <will@profian.com>